requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.2.2
django-redis==5.4.0
celery[redis]==5.3.6
mysqlclient==2.2.0